Fixed type of "count" parameter to match the declarations in io.h
git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@5470 1ca430b3-8271-4a6b-9b0e-3dcac98a8924

+++ nut/crt/read.c
@@ -72,7 +72,7 @@
  * \return The number of bytes read, which may be less than the number
  *         of bytes specified. A return value of -1 indicates an error.
  */
-int _read(int fd, void *buffer, unsigned int count)
+int _read(int fd, void *buffer, size_t count)
 {
     NUTFILE *fp = (NUTFILE *) ((uintptr_t) fd);
     NUTDEVICE *dev;
